split emission line editor
